test(cli): isolate terminal stream coverage
